// src/app/modules/profile/model/ChangePassword.ts
export class changePassword {
  Email: string = '';
  EmpId: number = 0;
  upwd: string = '';
  AddedBy: string = '';
  unm: string = '';
  IsActive: string = '';
  CityName: string = '';
  AddedOn: string = '';
  LastUpdatedOn: string = '';
  Action: string = '';
  NewPassword: string = '';
  oldpassword: string = '';
  confirmpassword: string = '';
  Encryptpwd: string = '';
  UserName:string='';
}